requirements.txt
imutils==0.5.3
numpy==1.18.4
opencv-python==4.2.0.34
opencv-contrib-python==4.2.0.34
pyzmq==19.0.1